+++ SevenSegment/SevenSegment.h
@@ -1,5 +1,5 @@
 /*
-* Display numbers from 0 to 15(0xA) on the SH5461AS
+* Display numbers from 0 to 15(0xF) on the SH5461AS
 * A->G: pin 2->pin 8
 * D1->D4: pin 9->pin 12
 *     A
@@ -36,7 +36,7 @@ private:
 public:
 	SevenSegment();
 	void displayDigit(byte digit, byte which = 0);
-	void displayNumber(long number);
+	void displayNumber(int number);
 	void setBlinking(byte which, bool blinking);
 	void setTurnedOn(byte which, bool turnedOn);
 	void disableBlinking();
